Have the filter directly modify whether the entry is featured
Before it was kind of a useless filter: whether to enable checking for
featured entry status, and if allowed to check the status, then use the
`is_starred` value. That was dumb. Now it passed the `is_starred` value
as whether the entry is featured, as expected.

Note: the filter name is no longer ideal, but it's what has been used
in the past. Ideally, the name should now reflect that it's directly
modifying the featured status, like
`gravityview_featured_entries_is_featured`.